+++ tests/not.rs
@@ -12,8 +12,7 @@ fn test_not() -> Result<(), Box<EvalAltResult>> {
     #[cfg(not(feature = "no_function"))]
     assert_eq!(engine.eval::<bool>("fn not(x) { !x } not(false)")?, true);
 
-    // TODO - do we allow stacking unary operators directly? e.g '!!!!!!!true'
-    assert_eq!(engine.eval::<bool>("!(!(!(!(true))))")?, true);
+    assert_eq!(engine.eval::<bool>("!!!!true")?, true);
 
     Ok(())
 }